+++ R/load_latest_forecasts_repo.R
@@ -14,10 +14,12 @@
 #' c('1 wk ahead cum death', '2 wk ahead cum death'). Defaults to all targets.
 #' 
 #' @return data frame with columns model, forecast_date, location, horizon,
-#' temporal_resolution, target_variable, target_end_date, type, quantile, value
+#' temporal_resolution, target_variable, target_end_date, type, quantile, value,
+#' location_name, population, geo_type, geo_value, abbreviation
 #'
 #' @export
-load_latest_forecasts_repo <- function(file_path, models, forecast_dates, locations, types, targets){
+load_latest_forecasts_repo <- function(file_path, models, forecast_dates, 
+                                       locations, types, targets){
   
   #validate file path to data-processed folder
   if (!dir.exists(file_path)){
@@ -122,7 +124,8 @@ load_latest_forecasts_repo <- function(file_path, models, forecast_dates, locati
     tidyr::separate(target, into=c("horizon","temporal_resolution","ahead","target_variable"),
                     remove = FALSE, extra = "merge") %>%
     dplyr::select(model, forecast_date, location, horizon, temporal_resolution, 
-                  target_variable, target_end_date, type, quantile, value)
+                  target_variable, target_end_date, type, quantile, value) %>%
+    dplyr::left_join(covidHubUtils::hub_locations, by=c("location" = "fips"))
 
   return(forecasts)