+++ R/plot_forecast.R
@@ -3,12 +3,16 @@
 #' @param forecast_data data frame with truth and forecasts from load_forecasts()
 #' @param truth_data optional data frame with forecasts in the format returned 
 #' by load_truth().
-#' @param model model_abbr specifying model to plot. Optional if there is only
-#' one model available in forecast data.
+#' @param models vector of strings specifying models to plot. 
+#' Default to all models in forecast_data.
 #' @param target_variable string specifying target type. It should be one of 
 #' "cum death", "inc case", "inc death"
-#' @param location string for fips code or 'US'. Optional if there is only one
-#' location available in forecast data.
+#' @param locations string for fips code or 'US'. 
+#' Default to all locations in forecast_data.
+#' @param facet interpretable facet option for ggplot
+#' @param facet_scales argument for scales in ggplot2::facet_wrap. Default to "fixed".
+#' @param forecast_dates date string vectors for forecast dates to plot. 
+#' Default to forecast_dates present in the data.
 #' @param intervals values indicating which central prediction interval levels 
 #' to plot. NULL means only plotting point forecasts.
 #' If not provided, it will default to c(.5, .8, .95).
@@ -21,6 +25,9 @@
 #' @param plot_truth boolean for showing truth data in plot. Default to FALSE.
 #' @param plot boolean for showing the plot. Default to TRUE.
 #' Currently supports "JHU","USAFacts", "NYTimes". Default to "JHU".
+#' @param fill_by_model boolean for specifying colors in plot.
+#' If TRUE, separate colors will be used for each model.
+#' If FALSE, only blues will be used for all models. Default to FALSE. 
 #' @param truth_as_of the plot includes the truth data that would have been 
 #' in real time as of the truth_as_of date (not using this parameter when truth data 
 #' is from github repo)
@@ -34,14 +41,18 @@
 #' @export
 plot_forecast <- function(forecast_data,
                           truth_data = NULL, 
-                          model,
+                          models,
                           target_variable,
-                          location,
+                          locations,
+                          facet = NULL,
+                          facet_scales = "fixed",
+                          forecast_dates,
                           intervals,
                           horizon,
                           truth_source = "JHU",
                           plot_truth = TRUE,
                           plot = TRUE,
+                          fill_by_model = FALSE,
                           truth_as_of = NULL, 
                           title = "default", 
                           show_caption = TRUE){
@@ -50,46 +61,36 @@ plot_forecast <- function(forecast_data,
   if(is.na(title))
     stop("title argument interpretable as a character.")
   
-  # optional model and location
-  if (length(unique(forecast_data$model)) == 1){
-    model = unique(forecast_data$model)
-  } else {
-    if (!missing(model)){
-      if (!(model %in% forecast_data$model)) {
-        stop("Error in plot_forecast: model is not available in forecast data.")
-      }
-    } else {
-      stop("Error in plot_forecast: Please select a model to plot.")
+  # optional models parameter. Default to all models in forecast_data
+  if (!missing(models)){
+    if (!all(models %in% forecast_data$model)) {
+      stop("Error in plot_forecast: Not all models are available in forecast data.")
     }
+  } else {
+    models <- unique(forecast_data$model)
   }
   
-  if (length(unique(forecast_data$location)) == 1){
-    location = unique(forecast_data$location)
-  } else {
-    if (!missing(location)){
-      if (!(location %in% forecast_data$location)) {
-        stop("Error in plot_forecast: location is not available in forecast data.")
-      }
-    } else {
-      stop("Error in plot_forecast: Please select a location to plot.")
-    }
+  # optional locations parameter. Default to all locations in forecast_data
+  if (missing(locations)){
+    locations <- unique(forecast_data$location)
   }
+  
+  # validate location fips code
+  all_valid_fips <- covidHubUtils::hub_locations$fips
+  
+  locations <- match.arg(locations, 
+                        choices = all_valid_fips, 
+                        several.ok = TRUE)
+  
+  if (!all(locations %in% forecast_data$location)){
+    stop("Error in plot_forecast: Not all locations are available in forecast_data.")
+  }
+  
   
   # validate truth_source
   truth_source <- match.arg(truth_source, 
                             choices = c("JHU","USAFacts", "NYTimes"), 
                             several.ok = FALSE)
-  
-  # validate location fips code
-  all_valid_fips <- covidHubUtils::hub_locations$fips
-  
-  location <- match.arg(location, 
-                        choices = all_valid_fips, 
-                        several.ok = FALSE)
-  
-  if (!location %in% forecast_data$location){
-    stop("Error in plot_forecast: location is not available in forecast_data.")
-  }
   
   # validate target_variable
   target_variable <- match.arg(target_variable, 
@@ -118,13 +119,24 @@ plot_forecast <- function(forecast_data,
         stop("Error in plot_forecast: Please provide a valid truth_source to plot.")
       }
       # check if truth_data has data from specified location
-      if (!(location %in% truth_data$location)){
+      if (!all(locations %in% truth_data$location)){
         stop("Error in plot_forecast: Please provide a valid location to plot.")
       }
       # check if truth_data has specified target variable
       if (!(target_variable %in% truth_data$target_variable)){
         stop("Error in plot_forecast: Please provide a valid target variable.")
       }
+    }
+  }
+  
+  
+  # validate forecast_dates
+  
+  if (missing(forecast_dates)){
+    forecast_dates <- unique(forecast_data$forecast_date)
+  } else {
+    if (!all(forecast_dates %in% forecast_data$forecast_date)){
+      stop ("Error in plot_forecast: Not all forecast_dates are available in forecast data.")
     }
   }
   
@@ -138,15 +150,25 @@ plot_forecast <- function(forecast_data,
     if ('NA' %in% lower_bounds){
       intervals = NULL
     } else {
-      intervals <- lapply(lower_bounds, function(l){
-        1-as.numeric(2*l)
-      })
-      
-      # for readability
-      if (all(c(.5, .8, .95) %in% intervals)){
-        intervals <- c(.5, .8, .95)
+      # only plot .95 interval if more than 5 models are selected
+      if (length(unique(models)) > 5){
+        intervals <- c(.95)
+      } else {
+        # generate intervals from lower bounds
+        intervals <- lapply(lower_bounds, function(l){
+          1-as.numeric(2*l)
+        })
+        
+        # for readability
+        if (all(c(.5, .8, .95) %in% intervals)){
+          intervals <- c(.5, .8, .95)
+        }
       }
     }
+  }
+  
+  if (!is.null(intervals) & length(unique(models)) > 5){
+    intervals <- c(.95)
   }
   
   # generate quantiles based on given intervals
@@ -156,36 +178,80 @@ plot_forecast <- function(forecast_data,
   }))
   
   
-  # prediction interval shades
-  if (!is.null(intervals)){
-    
-    colourCount = length(quantiles_to_plot)/2+1
+  # set colors
+  if (fill_by_model){
+    if (length(unique(models)) <= 5){
+      color_families <- c("Blues", "Oranges", "Greens", "Purples", "Reds")
+      colourCount <- length(quantiles_to_plot)/2+1
+      model_colors <- purrr::map(
+        color_families[1:length(unique(models))],
+        function(color_family){
+          getPalette = colorRampPalette(RColorBrewer::brewer.pal(4, color_family))
+          getPalette(colourCount)
+          }
+        )
+      
+      ribbon_colors <- RColorBrewer::brewer.pal(4, "Greys")[1:3]
+      forecast_colors <- unlist(lapply(model_colors, tail, n = 1))
+      interval_colors <- unlist(lapply(model_colors, head, n = colourCount-1))
+    } else {
+      # interpolate color pallets to more than 5 colors
+      colourCount <- length(quantiles_to_plot)/2
+      modelCount <- length(unique(models))
+      # use 8 instead of 9 to avoid black and grey shades
+      getPalette = colorRampPalette(RColorBrewer::brewer.pal(8, "Set1"))
+      model_colors <- getPalette(modelCount)
+      
+      ribbon_colors <- RColorBrewer::brewer.pal(4, "Greys")[1:3]
+      forecast_colors <- unlist(lapply(model_colors, tail, n = 1))
+      # create 95% PI color
+      interval_colors <- unlist(lapply(model_colors, function(color){
+        colorspace::adjust_transparency(color, 0.3)}
+        ))
+    }
+  } else {
+    # only use blue
+    colourCount = max(length(quantiles_to_plot)/2+1, 2)
     getPalette = colorRampPalette(RColorBrewer::brewer.pal(4, "Blues"))
     blues = getPalette(colourCount)
-  } else {
-    blues <- RColorBrewer::brewer.pal(n=4, "Blues")
+    forecast_colors <- rep(tail(blues,1), length(unique(models)))
+    interval_colors <- rep(blues[1:(length(blues)-1)],
+                           length(unique(models)))
+    ribbon_colors <- blues[1:(length(blues)-1)]
+  }
+  
+  if (!is.null(truth_as_of)){
+    warning("Warning in plot_forecast: truth_as_of is not used to load versioned truth data.
+            Will be available soon.")
   }
   
   # include truth from remote git hub repo by default
   # not using truth_as_of if we are loading truth from git hub repos
   plot_data = get_plot_forecast_data (forecast_data = forecast_data, 
                                       truth_data = truth_data,
-                                      model_to_plot = model,
+                                      models_to_plot = models,
+                                      forecast_dates_to_plot = as.Date(forecast_dates),
                                       horizons_to_plot = horizon,
                                       quantiles_to_plot = quantiles_to_plot,
-                                      location_to_plot = location,
+                                      locations_to_plot = locations,
                                       plot_truth = plot_truth,
                                       truth_source = truth_source,
                                       target_variable_to_plot = target_variable)
  
+
   # generate caption and full target variable
   if(show_caption){
     if(!is.null(truth_as_of)){
       caption <- paste0("source: ", truth_source," (observed data as of ",
-        as.Date(truth_as_of), "), ", model, " (forecasts)")
+                        as.Date(truth_as_of), "), ", 
+                        paste(models, collapse = ', '), " (forecasts)")
+      
+      caption <- paste(strwrap(caption, dev.size("px")[1]), collapse="\n")
     } else {
       caption <- paste0("source: ", truth_source," (observed data), ",
-        model," (forecasts)")
+                        paste(models, collapse = ', ')," (forecasts)")
+      
+      caption <- paste(strwrap(caption, dev.size("px")[1]), collapse="\n")
     }
   } else {
     caption <- NULL
@@ -201,50 +267,103 @@ plot_forecast <- function(forecast_data,
     full_target_variable = "Incident Hospitalizations"
   }
   
+  # split plot data
+  plot_data_forecast <- plot_data %>%
+    dplyr::filter(truth_forecast == "forecast")
+  
+  plot_data_truth <- plot_data %>%
+    dplyr::filter(!is.na(point),truth_forecast == "truth") %>%
+    dplyr::rename(truth_model = model) %>%
+    dplyr::select(-forecast_date)
+  
   # generate title if specified as "default", otherwise leave as is
   if(title == "default") {
-    title <- paste0("Weekly COVID-19 ", full_target_variable, " in ", 
-      location,": observed and forecasted")
+    title <- paste0("Weekly COVID-19 ", full_target_variable, 
+                    ": observed and forecasted")
+    subtitle <- paste0("Selected locations: ", 
+                       paste(unique(plot_data_forecast$location), collapse = ', ') )
   }
   if(title == "none") {
     title <- NULL
+    subtitle <- NULL
   }
-
   
-  graph <- ggplot2::ggplot(data = plot_data)
-    
+
+  # generate plot
+  graph <- ggplot2::ggplot(data = plot_data_forecast, ggplot2::aes(x= target_end_date))
+  
+  # plot selected prediction intervals
   if (!is.null(intervals)){
-    # plot all prediction intervals
     graph <- graph  +
-      ggplot2::geom_ribbon(data = plot_data %>%
+      ggplot2::geom_ribbon(data = plot_data_forecast %>%
                   dplyr::filter(type == "quantile"),
-                mapping = ggplot2::aes(x = target_end_date,
-                                       ymin=lower, ymax=upper,
-                                       fill=`Prediction Interval`)) +
-      
-      ggplot2::scale_fill_manual(values = blues[1:(length(blues)-1)])
-      
+                mapping = ggplot2::aes(ymin=lower, 
+                                       ymax=upper,
+                                       group = interaction(`Prediction Interval`, model, 
+                                                           location, forecast_date),
+                                       fill = interaction(`Prediction Interval`, model)),
+                show.legend=FALSE) +
+      ggplot2::scale_fill_manual(name = "Prediction Interval", 
+                                 values = interval_colors) +
+      # create a transparent layer with grey colors to get prediction interval legend
+      ggnewscale::new_scale_fill() +
+      ggplot2::geom_ribbon(data = plot_data_forecast %>%
+                             dplyr::filter(type == "quantile"),
+                           mapping = ggplot2::aes(ymin=lower, 
+                                                  ymax=upper, 
+                                                  fill = `Prediction Interval`), 
+                           alpha=0) +
+      ggplot2::scale_fill_manual(name = "Prediction Interval", 
+                                 values = ribbon_colors) +
+      # reset alpha in legend fill
+      ggplot2::guides(
+        fill = ggplot2::guide_legend(override.aes = list(alpha = 1)))
   }
   
   # plot point forecasts and truth 
   graph <- graph +
-    ggplot2::geom_line(data = plot_data %>%
+    #forecast
+    ggplot2::geom_line(data = plot_data_forecast %>%
                 dplyr::filter(!is.na(point)),
-              mapping = ggplot2::aes(x = target_end_date,
+              mapping = ggplot2::aes(x = target_end_date, 
                                      y = point, 
-                                     color = truth_forecast)) +
-    ggplot2::geom_point(data = plot_data %>%
+                                     group = interaction(model, location, forecast_date),
+                                     color = model)) +
+    ggplot2::geom_point(data = plot_data_forecast %>%
                  dplyr::filter(!is.na(point)),
                mapping = ggplot2::aes(x = target_end_date, 
                                       y = point, 
-                                      color = truth_forecast)) +
+                                      color = model)) +
     ggplot2::scale_color_manual(name = "Model", 
-                                label = unique(plot_data$model),
-                                values = c("truth" = "black",
-                                           "forecast" = tail(blues,1))) +
+                                values = forecast_colors) +
+    #truth
+    ggnewscale::new_scale_color() +
+    ggplot2::geom_line(data = plot_data_truth %>%
+                         dplyr::filter(!is.na(point)),
+                       mapping = ggplot2::aes(x = target_end_date, 
+                                              y = point, 
+                                              color = truth_model)) +
+    ggplot2::geom_point(data = plot_data_truth %>%
+                          dplyr::filter(!is.na(point)),
+                        mapping = ggplot2::aes(x = target_end_date, 
+                                               y = point, 
+                                               color = truth_model)) +
+    ggplot2::scale_color_manual(name = "Truth", values = "black")
+  
+  # add facets
+  if(!is.null(facet)){
+    graph <- graph + 
+      ggplot2::facet_wrap(facets = facet, scales = facet_scales, 
+                          labeller = ggplot2::label_wrap_gen(multi_line=FALSE))
+  
+  }
+  
+  # add labels, title and caption
+  graph <- graph + 
     ggplot2::scale_x_date(name = NULL, date_breaks="1 month", date_labels = "%b %d") +
     ggplot2::ylab(full_target_variable) +
-    ggplot2::labs(title = title ,
+    ggplot2::labs(title = title,
+                  subtitle = subtitle,
                   caption = caption)
   
   if (plot){