+++ NEWS.md
@@ -1,3 +1,19 @@
+## covidHubUtils 0.1.1
+
+This is a release focusing on new features in plotting functions. 
+
+### Feature updates
+- `plot_forecast()` now supports faceted plots of multiple models, locations and forecast dates for one target variable. 
+  
+  In `plot_forecast()`, `facet` and `facet_scales` are equivalent to `facets` and `scales` in `ggplot2::facet_wrap()`. `facet` takes facet formula, for example `facet = ~ model`. `facet_scales` are expecting the same values for `scales` in `ggplot2::facet_wrap()`, such as `"fixed"`, `"free_y"`, `"free_x"` or `"free"`.
+  
+  If `fill_by_model = TRUE`, each model will be represented by a unique color. If `fill_by model = FALSE`, all models and selected prediction intervals will be represented by blue colors.
+  
+  For simplicity, prediction interval legends will be grey in faceted plots. Morever, when the user selects more than 5 models, only 95% predicition interval is included. Otherwise, all selected prediction intervals will be plotted. 
+  
+### Package updates
+- There is no backwards compatibility due to argument changes in `plot_forecast()`.
+  
 ## covidHubUtils 0.1
 
 This is the first version of the package with a 0.x release.
@@ -6,6 +22,6 @@ This is the first version of the package with a 0.x release.
 - details on new features will be listed here for future updates
 - current key features include loading and plotting forecast and truth data
 
-### package updates
+### Package updates
 - details on other changes will be listed here for future updates
 - added initial author/contributor list